refactor: move logic to Surface component
